Rename the generated styles to css to prevent confusion
