+++ app/root.tsx
@@ -8,7 +8,7 @@ import {
   Scripts,
   ScrollRestoration,
 } from '@remix-run/react';
-import styles from './styles/app.css';
+import styles from './css/app.css';
 
 export const meta: MetaFunction = () => ({
   charset: 'utf-8',
