+++ app/routes/__main/settings/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '~/styles/routes/__main/settings.css';
+import styles from '~/css/routes/__main/settings.css';
 
 export function links() {
   return [{ rel: 'stylesheet', href: styles }];
